backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
pdfplumber>=0.10.0
PyPDF2>=3.0.0
python-docx>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.10
python-multipart>=0.0.6
pydantic-settings>=2.0.0
alembic>=1.12.0
